+++ zeroui/src/main/java/com/zero/zero_tools/zeroui/skin/ZeroSkinProvider.kt
@@ -55,7 +55,9 @@ public fun rememberZeroSkinFromMaterialTheme(): ZeroSkin {
                 mutedOutline = colorScheme.outlineVariant,
                 focusedOutline = colorScheme.primary,
                 errorOutline = colorScheme.error,
-                unknownContainer = Color(0x1A000000)
+                unknownContainer = Color(0x1A000000),
+                inverseContent = colorScheme.inverseOnSurface,
+                inverseContainer = colorScheme.inverseSurface
             ),
             typography = ZeroTypography(
                 title = typography.headlineMedium,